server/requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

# Flask Framework
Flask==3.0.3

# PostgreSQL Database
psycopg2-binary==2.9.9

# Dotenv
python-dotenv==1.0.1